Migrate MiniTest::Unit::TestCase class under the Test::Unit
